+++ Hospital ERP/src/main/java/hospital/dao/LoginDao.java
@@ -5,17 +5,22 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+import javax.sql.DataSource;
+
 import hospital.utils.ConnectionHelper;
 
 public class LoginDao {
 
+	public static LoginDao instance;
+	Connection conn;
+	PreparedStatement pstmt;
+	ResultSet rs;
+	
+	public LoginDao() {}
+
 	public int loginCheck(String h_id, String h_pwd) {
 
-		int ret = -1;
-		Connection conn = null;
-		PreparedStatement pstmt = null;
-		ResultSet rs = null;
-
+			int ret = -1;
 		try {
 			conn = ConnectionHelper.getConnection();
 			String sql = "SELECT COUNT(*) FROM HOSPITAL_ID WHERE H_ID = ? AND H_PWD = ?";
@@ -38,9 +43,6 @@ public class LoginDao {
 
 	public String ecode(String h_id) {
 
-		Connection conn = null;
-		PreparedStatement pstmt = null;
-		ResultSet rs = null;
 		String code = null;
 
 		try {
@@ -48,11 +50,11 @@ public class LoginDao {
 			String sql = "SELECT H_E_CODE FROM HOSPITAL_ID WHERE H_ID = ?";
 			pstmt = conn.prepareStatement(sql);
 			pstmt.setString(1, h_id);
-			
+
 			rs = pstmt.executeQuery();
-			while(rs.next()) {
-			code = rs.getString(1);
-			 }
+			while (rs.next()) {
+				code = rs.getString(1);
+			}
 		} catch (SQLException e) {
 			e.printStackTrace();
 		} finally {
@@ -64,9 +66,6 @@ public class LoginDao {
 
 	public String name(String e_code) {
 
-		Connection conn = null;
-		PreparedStatement pstmt = null;
-		ResultSet rs = null;
 		String name = null;
 
 		try {
@@ -90,11 +89,7 @@ public class LoginDao {
 
 	public String hospital(String h_name) {
 
-		Connection conn = null;
-		PreparedStatement pstmt = null;
-		ResultSet rs = null;
 		String id = null;
-
 		try {
 			conn = ConnectionHelper.getConnection();
 			String sql = "SELECT H_NAME FROM HOSPITAL_ID WHERE H_E_CODE = ?";
