+++ Hospital ERP/src/main/java/hospital/service/MemberIdCheckAction.java
@@ -15,7 +15,7 @@ public class MemberIdCheckAction implements Action {
 	@Override
 	public ActionForward execute(HttpServletRequest request, HttpServletResponse response) {
 
-		ActionForward forward = new ActionForward();
+		ActionForward forward = null;
 
 		String h_id = request.getParameter("h_id");
